Removed inbound_message_stream.rs
