+++ packages/tokio_krpc/src/recv_transport.rs
@@ -12,10 +12,14 @@ use crate::{
 use failure::ResultExt;
 use futures::{
     future,
+    stream,
     TryStream,
     TryStreamExt,
 };
-use krpc_encoding::Message;
+use krpc_encoding::{
+    Envelope,
+    Message,
+};
 use std::{
     self,
     net::SocketAddr,
@@ -60,7 +64,7 @@ impl RecvTransport {
     ) {
         let transactions = self.transactions.clone();
 
-        let query_stream = receive_inbound_messages(self.recv_half)
+        let query_stream = Self::receive_inbound_messages(self.recv_half)
             .map_ok(move |(envelope, from_addr)| match envelope.message_type {
                 Message::Response { .. } | Message::Error { .. } => {
                     transactions.handle_response(envelope)?;
@@ -79,4 +83,45 @@ impl RecvTransport {
             query_stream,
         )
     }
+
+    fn receive_inbound_messages(
+        recv_socket: UdpSocketRecvHalf,
+    ) -> impl TryStream<Ok = (Envelope, SocketAddr), Error = Error> {
+        let recv_buffer = [0 as u8; 1024];
+
+        stream::unfold((recv_socket, recv_buffer), |(recv_socket, recv_buffer)| {
+            receive_inbound_message_wrapper(recv_socket, recv_buffer)
+        })
+    }
+}
+
+async fn receive_inbound_message_wrapper(
+    mut recv_socket: UdpSocketRecvHalf,
+    mut recv_buffer: [u8; 1024],
+) -> Option<(
+    Result<(Envelope, SocketAddr)>,
+    (UdpSocketRecvHalf, [u8; 1024]),
+)> {
+    let result = receive_inbound_message(&mut recv_socket, &mut recv_buffer).await;
+
+    Some((result, (recv_socket, recv_buffer)))
+}
+
+async fn receive_inbound_message(
+    recv_socket: &mut UdpSocketRecvHalf,
+    recv_buffer: &mut [u8; 1024],
+) -> Result<(Envelope, SocketAddr)> {
+    let (size, from_addr) = recv_socket
+        .recv_from(recv_buffer)
+        .await
+        .map_err(|cause| ErrorKind::FailedToReceiveMessage { cause })?;
+
+    let envelope = Envelope::decode(&recv_buffer[..size]).with_context(|_| {
+        ErrorKind::InvalidInboundMessage {
+            from: from_addr,
+            message: recv_buffer[..size].to_vec(),
+        }
+    })?;
+
+    Ok((envelope, from_addr))
 }
